Allow specifying custom coordinates. (#94)
The upcoming version of flank is changing its maven coordinates. This
will allow us to test upcoming versions of flank using the current
version of fladle.

Also removed adding bintray repo. Adding a custom repo is not a good
idea for companies that use their own maven proxy for stability and
security reasons. This should also be a win for build speeds
once flank is uploaded to maven central.