Expose additionalTestApks in YamlConfigWriterTask itself (#291)
* Expose additionalTestApks in YamlConfigWriterTask itself

This allows for this input to be manually configured to source its value from a delegate task instead of the FladleConfig properties put in.

To achieve this easily, this also swaps the `@Nested` config handling for Kotlin class delegation instead. This functionally behaves the same but allows for easy overrides of properties like this. Future properties could be added here too.

* Don't fall back to additionalTestApks

Didn't do it before and gradle explodes if you try orElse-ing this anyway

* Merge config at task runtime instead